api/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
faiss-cpu==1.12.0
fastapi==0.116.2
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httptools==0.6.4
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.11.9
pydantic_core==2.33.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
